// frontend/src/types/proto/api/v1alpha/worksheet_service.ts
export interface Worksheet {
  /**
   * The name of the worksheet resource, generated by the server.
   * Canonical parent is project.
   * Format: worksheets/{worksheet}
   */
  name: string;
  /**
   * The project resource name.
   * Format: projects/{project}
   */
  project: string;
  /**
   * The database resource name.
   * Format: instances/{instance}/databases/{database}
   * If the database parent doesn't exist, the database field is empty.
   */
  database: string;
  /** The title of the worksheet. */
  title: string;
  /**
   * The creator of the Worksheet.
   * Format: users/{email}
   */
  creator: string;
  /** The create time of the worksheet. */
  createTime:
    | Timestamp
    | undefined;
  /** The last update time of the worksheet. */
  updateTime:
    | Timestamp
    | undefined;
  /**
   * The content of the worksheet.
   * By default, it will be cut off in SearchWorksheet() method. If it doesn't match the `content_size`, you can
   * use GetWorksheet() request to retrieve the full content.
   */
  content: Uint8Array;
  /** content_size is the full size of the content, may not match the size of the `content` field. */
  contentSize: Long;
  visibility: Worksheet_Visibility;
  /** starred indicates whether the worksheet is starred by the current authenticated user. */
  starred: boolean;
}

export enum Worksheet_Visibility {
  VISIBILITY_UNSPECIFIED = "VISIBILITY_UNSPECIFIED",
  /** VISIBILITY_PROJECT_READ - Read access in project scope, worksheet OWNER/DBA and project OWNER can read/write, other project members can read. */
  VISIBILITY_PROJECT_READ = "VISIBILITY_PROJECT_READ",
  /** VISIBILITY_PROJECT_WRITE - Write access in project scope, worksheet OWNER/DBA and all members in the project can write the worksheet. */
  VISIBILITY_PROJECT_WRITE = "VISIBILITY_PROJECT_WRITE",
  /** VISIBILITY_PRIVATE - Private, only worksheet OWNER can read/write. */
  VISIBILITY_PRIVATE = "VISIBILITY_PRIVATE",
  UNRECOGNIZED = "UNRECOGNIZED",
}